Refactor project structure: create huggingface module
This patch restructures the project structure by moving
huggingface-related classes and structures into the separate
directory `huggingface`.
