+++ include/metalchat/metalchat.h
@@ -9,7 +9,6 @@
 #include <metalchat/container.h>
 #include <metalchat/dtype.h>
 #include <metalchat/functional.h>
-#include <metalchat/huggingface.h>
 #include <metalchat/interpreter.h>
 #include <metalchat/kernel.h>
 #include <metalchat/kernel_thread.h>
