+++ test/test_huggingface.cc
@@ -5,7 +5,7 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers_floating_point.hpp>
 
-#include <metalchat/huggingface.h>
+#include <metalchat/huggingface/llama.h>
 
 #include "metalchat/testing.h"
